requirements.txt
torch==1.12.1+cu113 --extra-index-url https://download.pytorch.org/whl/cu113
torchvision==0.13.1+cu113 --extra-index-url https://download.pytorch.org/whl/cu113
torchaudio==0.12.1 --extra-index-url https://download.pytorch.org/whl/cu113
monai[nibabel, tqdm, einops]==1.2.0
monailabel==0.7.0
timm==0.6.13
tensorboard==2.13.0
ml-collections==0.1.1
scikit-learn==1.2.2
pandas==1.5.3
matplotlib==3.7.1
ray==2.4.0
torchsummaryX==1.3.0